Move `GSFetcher` to `internal/client/gcp` package
