+++ internal/client/fetcher.go
@@ -24,7 +24,6 @@ import (
 	"path"
 	"strings"
 
-	gcs "cloud.google.com/go/storage"
 	"github.com/transparency-dev/tessera/api/layout"
 	"k8s.io/klog/v2"
 )
@@ -122,56 +121,6 @@ func (f FileFetcher) ReadTile(_ context.Context, l, i uint64, p uint8) ([]byte, 
 
 func (f FileFetcher) ReadEntryBundle(_ context.Context, i uint64, p uint8) ([]byte, error) {
 	return os.ReadFile(path.Join(f.Root, ctEntriesPath(i, p)))
-}
-
-// NewGSFetcher creates a new GSFetcher for the Google Cloud Storage bucket, using
-// the provided GCS client.
-//
-// bucket should not contain any slash.
-// c may be nil, in which case a new GCS client will be used.
-func NewGSFetcher(ctx context.Context, bucket string, c *gcs.Client) (*GSFetcher, error) {
-	if c == nil {
-		var err error
-		c, err = gcs.NewClient(ctx, gcs.WithJSONReads())
-		if err != nil {
-			return nil, err
-		}
-	}
-	return &GSFetcher{
-		bucket: bucket,
-		c:      c,
-	}, nil
-}
-
-// GSFetcher knows how to fetch log artifacts from a Google Cloud Storage bucket.
-type GSFetcher struct {
-	bucket string
-	c      *gcs.Client
-}
-
-func (f GSFetcher) fetch(ctx context.Context, p string) ([]byte, error) {
-	r, err := f.c.Bucket(f.bucket).Object(p).NewReader(ctx)
-	if err != nil {
-		return nil, fmt.Errorf("getObject: failed to create reader for object %q in bucket %q: %w", p, f.bucket, err)
-	}
-
-	d, err := io.ReadAll(r)
-	if err != nil {
-		return nil, fmt.Errorf("failed to read %q: %v", p, err)
-	}
-	return d, r.Close()
-}
-
-func (f GSFetcher) ReadCheckpoint(ctx context.Context) ([]byte, error) {
-	return f.fetch(ctx, layout.CheckpointPath)
-}
-
-func (f GSFetcher) ReadTile(ctx context.Context, l, i uint64, p uint8) ([]byte, error) {
-	return f.fetch(ctx, layout.TilePath(l, i, p))
-}
-
-func (f GSFetcher) ReadEntryBundle(ctx context.Context, i uint64, p uint8) ([]byte, error) {
-	return f.fetch(ctx, ctEntriesPath(i, p))
 }
 
 func ctEntriesPath(n uint64, p uint8) string {
